refactor Move domain factory to domain layer
